+++ scripts/debian/custom_provisioners/helloworld/A_base.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+echo '##############################################'
+echo "Starting ${0}.."
+set -x
+
+echo "Hello world"