+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+[Unreleased] ([unreleased])
+---------------------------
+* add custom provisioner helloworld and salt_arbe_git
+
 v0.4 (2015-03-20)
 -----------------
 * added CHANGELOG